LayoutTests/fast/attachment/attachment-disabled-dom.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This tests that attachments have an instance type of HTMLUnknownElement if attachments are disabled.");

var attachment = document.createElement("attachment");
shouldBeType("attachment", "HTMLUnknownElement");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
