LayoutTests/fast/attachment/attachment-disabled-dom-expected.txt

This tests that attachments have an instance type of HTMLUnknownElement if attachments are disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attachment is an instance of HTMLUnknownElement
PASS successfullyParsed is true

TEST COMPLETE
